pyproject.toml
[tool.poetry]
name = "hearts"
version = "0.1.0"
description = ""
authors = ["robot <robotnaoborot@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.5"}
beanie = "^1.27.0"
pydantic-settings = "^2.6.1"
uvicorn = "^0.32.0"
python-decouple = "^3.8"
aiogram = "^3.15.0"


[tool.poetry.group.dev.dependencies]
poetry-plugin-export = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-auto-export]
output = "requirements.txt"
without_hashes = true
without = ["dev"]